requirements.txt
pandas>=2.0.0
google-genai>=1.5.0
litellm>=1.63.0
python-dotenv>=1.0.1